Add FELightingSoftwareApplierInlines.h https://bugs.webkit.org/show_bug.cgi?id=267602
Reviewed by Said Abou-Hallawa.

`FELightingSoftwareApplier::LightingData` struct declares a number of
`inline` methods which were defined in `FELightingSoftwareApplier.cpp`.
However usage outside of the file happened in 272873@main. Unified
builds continued to function because the `.cpp` files using it ended up
in the same compilation unit, but non-unified builds failed with a
linker error.

Move the inlined `FELightingSoftwareApplier::LightingData` methods to
a new file `FELightingSoftwareApplierInlines.h` and include them when
applicable.

* Source/WebCore/platform/graphics/filters/software/FELightingSoftwareApplier.cpp:
* Source/WebCore/platform/graphics/filters/software/FELightingSoftwareApplierInlines.h: Copied from Source\WebCore\platform\graphics\filters\software\FELightingSoftwareApplier.cpp.
* Source/WebCore/platform/graphics/filters/software/FELightingSoftwareParallelApplier.cpp:

Canonical link: https://commits.webkit.org/273112@main
